[#32]feat: 프로필 푸터 컴포넌트
